Renames package
