+++ cmd/banksaurus/command/report/command.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/luistm/banksaurus/next/adapter/cgdcsv"
-	"github.com/luistm/banksaurus/next/adapter/transactionpresenter"
+	"github.com/luistm/banksaurus/next/adapter/presenterlisttransactions"
 	"github.com/luistm/banksaurus/next/listtransactions"
 	"github.com/luistm/banksaurus/next/listtransactionsgrouped"
 )
@@ -47,7 +47,7 @@ func (rc *Command) Execute(arguments map[string]interface{}) error {
 		return err
 	}
 
-	p, err := transactionpresenter.NewPresenter()
+	p, err := presenterlisttransactions.NewPresenter()
 	if err != nil {
 		return err
 	}
