+++ next/adapter/presenterlisttransactions/doc.go
@@ -1,0 +1,4 @@
+// Package presenterlisttransactions repackages the interactor data
+// into a view model
+
+package presenterlisttransactions
