+++ next/adapter/presenterlisttransactions/interfaces.go
@@ -1,4 +1,4 @@
-package transactionpresenter
+package presenterlisttransactions
 
 import "io"
