+++ next/adapter/presenterlisttransactions/presenter.go
@@ -1,4 +1,4 @@
-package transactionpresenter
+package presenterlisttransactions
 
 import (
 	"errors"
