--- next/adapter/transactionpresenter/doc.go
@@ -1,4 +1,0 @@
-// Package transactionpresenter repackages the interactor data
-// into a view model
-
-package transactionpresenter
